Extract blame function
